moved inside_sdp code into inside_sdp_images
